Move the remote package from pkg to model
